Fold configuration into processor
The processor is so thin now, there's no sense having a separate mixin
for configuration.
