--- aspen/processor.py
@@ -1,34 +1,0 @@
-"""
-aspen.processor
-+++++++++++++++
-"""
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
-from algorithm import Algorithm
-from .configuration import Configurable
-
-
-class Processor(Configurable):
-    """Model a processor of simplates.
-    """
-
-    def __init__(self, **kwargs):
-        """Takes configuration in kwargs.
-        """
-        self.algorithm = Algorithm.from_dotted_name('aspen.algorithm')
-        self.configure(**kwargs)
-
-
-    def process(self, path, querystring, accept_header, raise_immediately=None, return_after=None):
-        """Given a WSGI environ, return a state dict.
-        """
-        return self.algorithm.run( processor=self
-                                 , path=path
-                                 , querystring=querystring
-                                 , accept_header=accept_header
-                                 , _raise_immediately=raise_immediately
-                                 , _return_after=return_after
-                                  )
